refactor: share fake Codex runtime fixture
